--- src/module/user/user.entity.ts
@@ -1,19 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
-
-@Entity()
-export class User {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  name: string;
-
-  @Column({ unique: true }) // 중복 허용 안함.
-  email: string;
-
-  @Column()
-  password: string;
-
-  @Column({ default: '' })
-  introduce: string;
-}
